+++ examples/docs/remedy.config.ts
@@ -7,7 +7,7 @@ export default {
         "tailwind",
         "refresh",
         "static",
-        "partial",
+        "template",
         markdown({
             theme: "themes/markdown.yml",
             languages: ["ts", "sh"],