+++ packages/server/src/cli.ts
@@ -1,50 +1,55 @@
 import { info } from "@niarada/remedy-common";
-import { Option, program } from "commander";
+import { Command, Option } from "commander";
 import { create } from "./create";
 import { serve } from "./server";
 
 const pkg = require("../package.json");
 
 export function cli() {
-    const options = program.parseOptions(process.argv);
-    if (!options.operands[2]) {
-        process.argv.push("serve");
+    const argv = structuredClone(process.argv);
+    const program = createProgram();
+    if (!["-h", "--help", "--version", "-V", ...program.commands.map((c) => c.name())].includes(argv[2])) {
+        argv.splice(2, 0, "serve");
     }
-    program.parse();
+    program.parse(argv);
 }
 
-program
-    .name("remedy")
-    .version(pkg.version)
-    .description(`remedy hypermedia server v${pkg.version}
+const description = `remedy hypermedia server v${pkg.version}
 
 If no arguments are provided, runs the server in the current
 directory, serving the specified or configured public folder,
 or the current directory.
-`);
-program.addOption(new Option("--is-hot").hideHelp());
+`;
 
-program
-    .command("new")
-    .description("Create a new remedy project")
-    .argument("[path]", "The path to create the project in")
-    .action((path) => {
-        create({ path });
-    });
+function createProgram() {
+    const program = new Command();
+    program.name("remedy").version(pkg.version).description(description);
+    program.addOption(new Option("--is-hot").hideHelp());
 
-program
-    .command("serve")
-    .description("Start the hypermedia server")
-    .argument("[public]", "The public directory to serve from")
-    .option("-p, --port <number>", "The port to listen on")
-    .option("-P, --public <path>", "The public directory to serve from")
-    .option("-C, --create-config", "Create a config file")
-    .action((path: string, options, command) => {
-        if (path) {
-            options.public = path;
-        }
-        serve(options);
-        if (command.parent.opts().isHot) {
-            info("server", "running hot");
-        }
-    });
+    program
+        .command("new")
+        .description("Create a new remedy project")
+        .argument("[path]", "The path to create the project in")
+        .action((path) => {
+            create({ path });
+        });
+
+    program
+        .command("serve")
+        .description("Start the hypermedia server")
+        .argument("[public]", "The public directory to serve from")
+        .option("-p, --port <number>", "The port to listen on")
+        .option("-P, --public <path>", "The public directory to serve from")
+        .option("-C, --create-config", "Create a config file")
+        .action((path: string, options, command) => {
+            if (path) {
+                options.public = path;
+            }
+            serve(options);
+            if (command.parent.opts().isHot) {
+                info("server", "running hot");
+            }
+        });
+
+    return program;
+}